Refactor API chat handling into dedicated `chat` package
Moved `handleAsk` logic to `internal/handler/chat/chat.go` for better modularity and maintainability. Improved code structure by encapsulating chat-related functionalities, reducing duplication, and enhancing readability.
